* amar-src/amarchiver.c: Cast amar-size result ot (long long)
git-svn-id: https://svn.code.sf.net/p/amanda/code/amanda/trunk@6838 a8d146d6-cc15-0410-8900-af154a0219e0

+++ amar-src/amarchiver.c
@@ -347,7 +347,7 @@ do_list(
 	    /* one of the callbacks already printed an error message */
 	    exit(1);
     }
-printf("size: %lld\n", amar_size(archive));
+    printf("size: %lld\n", (long long)amar_size(archive));
     amar_close(archive, NULL);
 }
